# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bloom Learning App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  /** @minLength 6 */
  password: string;
  /**
   * @minLength 1
   * @maxLength 30
   */
  displayName: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  coins: number;
  streakCount: number;
  lastActivityDate: string | null;
  lastGiftDate: string | null;
  avatarData: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateUserBody {
  /**
   * @minLength 1
   * @maxLength 30
   */
  displayName?: string;
  avatarData?: string;
}

export interface EarnCoinsBody {
  /** @minimum 0 */
  amount: number;
}

export interface CoinsResponse {
  coins: number;
}

export interface ClaimGiftResponse {
  coins: number;
  message: string;
  newBalance: number;
}

export interface StreakResponse {
  streakCount: number;
  isNewDay: boolean;
  message: string;
}

export type AnimalRarity = (typeof AnimalRarity)[keyof typeof AnimalRarity];

export const AnimalRarity = {
  common: "common",
  uncommon: "uncommon",
  rare: "rare",
  epic: "epic",
  legendary: "legendary",
} as const;

export interface Animal {
  id: number;
  name: string;
  emoji: string;
  description: string;
  cost: number;
  rarity: AnimalRarity;
}

export interface PurchaseAnimalBody {
  animalId: number;
}

export interface PurchaseAnimalResponse {
  animal: Animal;
  newBalance: number;
}

export interface LeaderboardEntry {
  userId: number;
  displayName: string;
  avatarData: string | null;
  coins: number;
  streakCount: number;
  animalCount: number;
  animals: Animal[];
}

export type GenerateLessonBodySubject =
  (typeof GenerateLessonBodySubject)[keyof typeof GenerateLessonBodySubject];

export const GenerateLessonBodySubject = {
  math: "math",
  french: "french",
  spanish: "spanish",
  maltese: "maltese",
  italian: "italian",
  grammar: "grammar",
  history: "history",
  geography: "geography",
} as const;

export interface GenerateLessonBody {
  subject: GenerateLessonBodySubject;
  exerciseType: string;
  /**
   * @minimum 1
   * @maximum 12
   */
  level: number;
  language?: string | null;
}

export type LessonQuestionType =
  (typeof LessonQuestionType)[keyof typeof LessonQuestionType];

export const LessonQuestionType = {
  multiple_choice: "multiple_choice",
  fill_blank: "fill_blank",
  match: "match",
  write: "write",
  speak: "speak",
} as const;

export interface MatchPair {
  left: string;
  right: string;
}

export interface LessonQuestion {
  id: string;
  question: string;
  type: LessonQuestionType;
  options: string[] | null;
  correctAnswer: string;
  pairs: MatchPair[] | null;
}

export interface Lesson {
  id: string;
  subject: string;
  exerciseType: string;
  level: number;
  title: string;
  content: string;
  questions: LessonQuestion[];
  audioText: string | null;
}

export type TTSBodyVoice =
  | (typeof TTSBodyVoice)[keyof typeof TTSBodyVoice]
  | null;

export const TTSBodyVoice = {
  alloy: "alloy",
  echo: "echo",
  fable: "fable",
  onyx: "onyx",
  nova: "nova",
  shimmer: "shimmer",
} as const;

export interface TTSBody {
  text: string;
  voice?: TTSBodyVoice;
}

export interface TTSResponse {
  /** Base64-encoded MP3 audio */
  audio: string;
}

export interface TranscribeBody {
  /** Base64-encoded audio data */
  audio: string;
}

export interface TranscribeResponse {
  text: string;
}

export interface DailyChallenge {
  id: number;
  date: string;
  subject: string;
  title: string;
  description: string;
  coinReward: number;
  completed: boolean;
  questions: LessonQuestion[];
}

export interface CompleteChallengeBody {
  challengeId: number;
  answers: string[];
}

export interface CompleteChallengeResponse {
  coinsEarned: number;
  newBalance: number;
  score: number;
  passed: boolean;
}

export interface SubjectProgress {
  subject: string;
  currentLevel: number;
  highestScore: number;
  lessonsCompleted: number;
  lastActivityAt: string | null;
}

export interface SaveProgressBody {
  subject: string;
  level: number;
  score: number;
  exerciseType: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface SendOpenaiVoiceMessageBody {
  /** Base64-encoded audio data */
  audio: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "512x512": "512x512",
  "256x256": "256x256",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}