# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bloom Learning App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /animals/owned:
    get:
      operationId: getOwnedAnimals
      tags: [animals]
      summary: Get user's owned animals
      responses:
        "200":
          description: User's animals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Animal"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /animals/purchase:
    post:
      operationId: purchaseAnimal
      tags: [animals]
      summary: Purchase an animal with coins
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseAnimalBody"
      responses:
        "200":
          description: Purchase successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseAnimalResponse"
        "400":
          description: Not enough coins or already owned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /daily-challenge:
    get:
      operationId: getDailyChallenge
      tags: [dailyChallenge]
      summary: Get today's daily challenge
      responses:
        "200":
          description: Daily challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyChallenge"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /daily-challenge/complete:
    post:
      operationId: completeDailyChallenge
      tags: [dailyChallenge]
      summary: Complete the daily challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteChallengeBody"
      responses:
        "200":
          description: Completion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteChallengeResponse"
        "400":
          description: Already completed or wrong answers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    RegisterBody:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        password:
          type: string
          minLength: 6
        displayName:
          type: string
          minLength: 1
          maxLength: 30
      required:
        - username
        - password
        - displayName
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        coins:
          type: integer
        streakCount:
          type: integer
        lastActivityDate:
          type: string
          nullable: true
        lastGiftDate:
          type: string
          nullable: true
        avatarData:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - displayName
        - coins
        - streakCount
        - lastActivityDate
        - lastGiftDate
        - avatarData
        - createdAt
...
    Animal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        emoji:
          type: string
        description:
          type: string
        cost:
          type: integer
        rarity:
          type: string
          enum: [common, uncommon, rare, epic, legendary]
      required:
        - id
        - name
        - emoji
        - description
        - cost
        - rarity
    PurchaseAnimalBody:
      type: object
      properties:
        animalId:
          type: integer
      required:
        - animalId
    PurchaseAnimalResponse:
      type: object
      properties:
        animal:
          $ref: "#/components/schemas/Animal"
        newBalance:
          type: integer
      required:
        - animal
        - newBalance
...
    LessonQuestion:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        type:
          type: string
          enum: [multiple_choice, fill_blank, match, write, speak]
        options:
          type: array
          items:
            type: string
          nullable: true
        correctAnswer:
          type: string
        pairs:
          type: array
          items:
            $ref: "#/components/schemas/MatchPair"
          nullable: true
      required:
        - id
        - question
        - type
        - options
        - correctAnswer
        - pairs
    MatchPair:
      type: object
      properties:
        left:
          type: string
        right:
          type: string
      required:
        - left
        - right
...
    DailyChallenge:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        subject:
          type: string
        title:
          type: string
        description:
          type: string
        coinReward:
          type: integer
        completed:
          type: boolean
        questions:
          type: array
          items:
            $ref: "#/components/schemas/LessonQuestion"
      required:
        - id
        - date
        - subject
        - title
        - description
        - coinReward
        - completed
        - questions
    CompleteChallengeBody:
      type: object
      properties:
        challengeId:
          type: integer
        answers:
          type: array
          items:
            type: string
      required:
        - challengeId
        - answers
    CompleteChallengeResponse:
      type: object
      properties:
        coinsEarned:
          type: integer
        newBalance:
          type: integer
        score:
          type: integer
        passed:
          type: boolean
      required:
        - coinsEarned
        - newBalance
        - score
        - passed
...