# lib/api-zod/src/generated/types/completeChallengeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bloom Learning App API
 * OpenAPI spec version: 0.1.0
 */

export interface CompleteChallengeResponse {
  coinsEarned: number;
  newBalance: number;
  score: number;
  passed: boolean;
}