# lib/api-zod/src/generated/types/dailyChallenge.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bloom Learning App API
 * OpenAPI spec version: 0.1.0
 */
import type { LessonQuestion } from "./lessonQuestion";

export interface DailyChallenge {
  id: number;
  date: string;
  subject: string;
  title: string;
  description: string;
  coinReward: number;
  completed: boolean;
  questions: LessonQuestion[];
}