# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bloom Learning App API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  displayName: string;
  coins: number;
  streakCount: number;
  lastActivityDate: string | null;
  lastGiftDate: string | null;
  avatarData: string | null;
  createdAt: Date;
}